add mock component into popap
