+++ src/mock/mockDataComments.js
@@ -1,0 +1,27 @@
+
+export const mockDataComments = {
+  commentsEmoji: [
+    `./images/emoji/smile.png`,
+    `./images/emoji/sleeping.png`,
+    `./images/emoji/puke.png`,
+    `./images/emoji/angry.png`
+  ],
+  commentsText: [
+    `Interesting setting and a good cast`,
+    `Booooooooooring`,
+    `Very very old. Meh`,
+    `Almost two hours? Seriously?`
+  ],
+  commentsAuthor: [
+    `Tim Macoveev`,
+    `John Doe`,
+    `Liz Bakner`,
+    `John Smith`
+  ],
+  commentsDay: [
+    `2019/12/31 23:59`,
+    `2 days ago`,
+    `week ago`,
+    `Today`
+  ]
+};
