+++ src/mock/mockDataFilms.js
@@ -1,0 +1,67 @@
+export const mockDataFilms = {
+  titles: [
+    `The Dance of Life`,
+    `Sagebrush Trail`,
+    `The Man with the Golden Arm`,
+    `Santa Claus Conquers the Martians`,
+    `Popeye the Sailor Meets Sindbad the Sailor`
+  ],
+  posters: [
+    `./images/posters/the-dance-of-life.jpg`,
+    `./images/posters/sagebrush-trail.jpg`,
+    `./images/posters/the-man-with-the-golden-arm.jpg`,
+    `./images/posters/santa-claus-conquers-the-martians.jpg`,
+    `./images/posters/popeye-meets-sinbad.png`
+  ],
+  directors: [
+    `Anthony Mann`,
+    `George Lukas`,
+    `Stiven Spielberg`,
+    `Peter Jackson`,
+    `Luc Besson`
+  ],
+  writers: [
+    `Anne Wigton`,
+    `Heinz Herald`,
+    `Richard Weil`,
+    `Rian Johnson`,
+    `Todd Phillips`
+  ],
+  actors: [
+    `Erich von Stroheim`,
+    `Mary Beth Hughes`,
+    `Dan Duryea`,
+    `Harrison Ford`,
+    `Tommy Lee Jones`
+  ],
+  country: [
+    `USA`,
+    `Britain`,
+    `Germany`,
+    `China`,
+    `France`
+  ],
+  descriptions: [
+    `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
+    `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget.`,
+    `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra.`,
+    `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`,
+    `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.`
+  ],
+  durationTime: [
+    `1h 55m`,
+    `54m`,
+    `1h 59m`,
+    `1h 21m`,
+    `16m`
+  ],
+  genre: [
+    `Musical`,
+    `Western`,
+    `Drama`,
+    `Comedy`,
+    `Cartoon`,
+    `Mystery`,
+    `Romance`
+  ]
+}
